requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
langchain==0.3.7
langchain-core==0.3.18
langchain-openai==0.2.8
python-dotenv==1.0.1
